Rename all embedders to embedder_<name>.py, add is_default property
The previous convention used a special embedder.py for each media type's
default and embedder_<variant>.py for alternatives, with the default
selected by alphabetical filename order. That made the default implicit
and required the default file to be alphabetically first.

Now every embedder lives in its own embedder_<name>.py file, and the
default is declared explicitly via the is_default property on
MediaEmbedder (default False, overridden by CLAP, E5, X-CLIP, and SigLIP).
embedders_for_type() sorts defaults first, so callers using
embedders_for_type(t)[0] continue to receive the default.

Renames:
  audio/embedder.py -> audio/embedder_clap.py
  text/embedder.py  -> text/embedder_e5.py
  video/embedder.py -> video/embedder_xclip.py
(image/embedder_siglip.py already followed the convention.)
