+++ vtsearch/media/audio/embedder_clap.py
@@ -45,6 +45,10 @@ class AudioClapEmbedder(MediaEmbedder):
     def media_type_id(self) -> str:
         return "audio"
 
+    @property
+    def is_default(self) -> bool:
+        return True
+
     # ------------------------------------------------------------------
     # Model lifecycle
     # ------------------------------------------------------------------
