+++ vtsearch/media/text/embedder_e5.py
@@ -46,6 +46,10 @@ class TextE5Embedder(MediaEmbedder):
     def media_type_id(self) -> str:
         return "text"
 
+    @property
+    def is_default(self) -> bool:
+        return True
+
     # ------------------------------------------------------------------
     # Model lifecycle
     # ------------------------------------------------------------------
